Writing the IUI's CLI configuration file to iui dir if on Windows
This alters the default behavior of the CLI ever so slightly.  Adding
this file will prompt for user input when the user starts the InVEST
application without specifying a model.  This is particularly useful
on Windows, when the user double-clicks the EXE.  This way, the cmd
window will prompt for user input instead of just flashing open and
closing.