Merging master into release/3.3.0
Making this merge into release/3.3.0 is unorthodox and shouldn't need to
happen as part of a normal release.  I'm doing it here so I can have the
option to make a few more commits if needed before I merge back into
master for the final release tag.

This merge is really complex because hg has no history of
changes between the two branches being merged.  This means that I did a
lot of manual reverting to be able to make this happen, and I just hope
that it didn't mess anything up.  One more build and test before
tagging.  RE#3511