Improving two issues with CV:
The first is a correction to an assert statement.  Now, instead of
asserting that there is only one value less than or equal to 0, I just
set all values below 0 to 0.

The second is a correction for an indexError when a shoreline detection
routine tries to extend off the edge of an array.  If this happens, I
now skip the shoreline segment.

RE:#3624