Cleaning up and documenting Executor.
This commit also removes the unused logfile and tempdir parameters to
Executor.__init__(), as both of these can be handled from within the
callable being passed as needed by the particular application.  All the
Executor should do is run the callable, log an exception and emit a Qt
signal when complete.  RE:#3610