Refactoring again, back to a ValidationContext object.
While I've found that I can use functions, closures, et cetera to
provide the same sort of functionality (and sometimes with a couple
fewer lines), I've not liked the hidden behavior of the functions.  This
object-oriented approach makes it very clear that we are preserving the
state of the validation warnings through the validate function, but
because of the design of the object, the programmer is not required to
use it.  So long as validate() adheres to the validation API (and we
have the validation.validator decorator to help with that), the
developer should be good to go.

That being said, I do think that the ValidationContext object makes some
things easier :)

RE:#3610  RE:#3597